requirements.txt

viam-sdk==0.50.0

typing-extensions
